83000 refactoring (#61)
* WIP: unprepared changes

* WIP: untested- latest working logic app

* WIP: hard coded parameters

* Logic app update and script changes for cotnrol states

* Latest changes to logic app

* Fixes

+++ src/Dfe.CdcEventApi.Application/LoadProcessor.cs
@@ -194,7 +194,7 @@
         /// </returns>
         public async Task UpdateLoadStatusAsync(
             DateTime runIdentifier,
-            short status,
+            int status,
             CancellationToken cancellationToken)
         {
             this.loggerProvider.Debug($"Updating the load status...");
